+++ arm.urdf
@@ -7,11 +7,36 @@
     <link name="elbow_wrist_link"></link>
     <link name="wrist_palm_link"></link>
 
-    <joint name="base" type="continuous">
+    <joint name="base" type="revolution">
         <parent link="world_base_link"/>
         <child link="base_shoulder_link"/>
         <origin xyz="5 3 0" rpy="0 0 0"/>
         <axis xyz="-0.9 0.15 0"/>
     </joint>
 
+    <joint name="base" type="revolute">
+        <parent link="world_base_link"/>
+        <child link="base_shoulder_link"/>
+
+        <origin xyz="5 3 0" rpy="0 0 0"/>
+        <axis xyz="-0.9 0.15 0"/>
+    </joint>
+    <joint name="base" type="revolution">
+        <parent link="world_base_link"/>
+        <child link="base_shoulder_link"/>
+        <origin xyz="5 3 0" rpy="0 0 0"/>
+        <axis xyz="-0.9 0.15 0"/>
+    </joint>
+    <joint name="base" type="revolution">
+        <parent link="world_base_link"/>
+        <child link="base_shoulder_link"/>
+        <origin xyz="5 3 0" rpy="0 0 0"/>
+        <axis xyz="-0.9 0.15 0"/>
+    </joint>
+    <joint name="base" type="revolution">
+        <parent link="world_base_link"/>
+        <child link="base_shoulder_link"/>
+        <origin xyz="5 3 0" rpy="0 0 0"/>
+        <axis xyz="-0.9 0.15 0"/>
+    </joint>
 </robot>
