+++ arm_controller/chains/urdf_joint.py
@@ -1,0 +1,20 @@
+from arm_controller.chains.py_segment import PySegment
+
+
+class URDFJoint:
+    """
+    Our representation of URDF Joints for URDF-Compliant Solvers
+    """
+    def __init__(self, ps: PySegment):
+        if ps.joint_rot == 'X':
+            axis = [1, 0, 0]
+        elif ps.joint_rot == 'Y':
+            axis = [0, 1, 0]
+        elif ps.joint_rot == 'Z':
+            axis = [0, 0, 1]
+        else:
+            axis = [0, 0, 0]
+        self.name = ps.function
+        self.origin_xyz = ps.translation
+        self.origin_rpy = ps.rotation
+        self.axis_xyz = axis
