+++ arm_controller/chains/urdf_link.py
@@ -1,0 +1,8 @@
+from arm_controller.chains.py_segment import PySegment
+
+
+class URDFLink:
+    """
+    Our representation of URDF Links for URDF-Compliant Solvers
+    """
+
